+++ sw/params.c
@@ -30,13 +30,13 @@ uint32_t param_sample_rate(uint32_t desired_sample_rate_hz, uint16_t ovs_bits, u
 	*t_div = XMEGA_CLKDIV_8_gc;
 	*t_ovf = (ovf-1); // -1 is extremely important! this is for overflow
 
-	float filpot_r = 1.0 / ((desired_sample_rate_hz/2.0)*2.0*M_PI*FILCAP_F);
+	double filpot_r = 1.0 / ((desired_sample_rate_hz/2.0)*2.0*M_PI*FILCAP_F);
 	*filpot_pos = (filpot_r / FILPOT_OHM) * 1024;
 
 	return actual_sample_rate_hz;
 } //}}}
 
-float param_gain(uint32_t desired_gain, uint16_t* amppot_pos) //{{{
+double param_gain(uint32_t desired_gain, uint16_t* amppot_pos) //{{{
 {
 	if (desired_gain <= 1)
 	{
@@ -51,5 +51,5 @@ float param_gain(uint32_t desired_gain, uint16_t* amppot_pos) //{{{
 		fprintf(stderr, "warning: gain out of range, reverting to default\n");
 		*amppot_pos = AMPPOT_POS_DEFAULT;
 	}
-	return (100000.0 / ((((float)*amppot_pos)/1024.0) * AMPPOT_OHM)) + 1.0;
+	return (100000.0 / ((((double)*amppot_pos)/1024.0) * AMPPOT_OHM)) + 1.0;
 } //}}}
